Generate overall marker expressions across Coereba clusters,

Generate overall marker expressions across Coereba clusters,

Usage

Coereba_MarkerExpressions(
  data,
  binary,
  panel,
  starter,
  returnType = "All",
  CombinatorialArgs = NULL
)

Arguments

data

The data.frame of clusters vs individual specimens, with the ratio values

binary

The data.frame of markers vs clusters, with 0 and 1 values

panel

A .csv or data.frame containing Fluorophore and Marker columns of your panel markers.

starter

A string containing the starting characters for all Coereba cluster names

returnType

Either "All" or "Combinatorial". Default is ALL.

CombinatorialArgs

When returnType Combinatorial, the two fluorophores to create quadrants for.

Value

returns a data.frame

Examples

File_Location <- system.file("extdata", package = "Coereba")
panelPath <- file.path(File_Location, "ILTPanelTetramer.csv")
panelData <- read.csv(panelPath, check.names=FALSE)
binaryPath <- file.path(File_Location, "HeatmapExample.csv")
binaryData <- read.csv(binaryPath, check.names=FALSE)
dataPath <- file.path(File_Location, "ReadyFileExample.csv")
dataData <- read.csv(dataPath, check.names=FALSE)

All <- Coereba_MarkerExpressions(data=dataData, binary=binaryData,
 panel=panelData, starter="SparkBlue550")

Memory <- Coereba_MarkerExpressions(data=dataData, binary=binaryData,
 panel=panelData, starter="SparkBlue550", returnType = "Combinatorial",
 CombinatorialArgs=c("BV510", "APC-Fire 750"))