Normality Test followed by t-test/Anova

Normality Test followed by t-test/Anova

Usage

Utility_Stats(
  data,
  var,
  myfactor,
  normality = NULL,
  specifiedNormality = NULL,
  correction = "none",
  override = 0.05,
  returnType = "stats"
)

Arguments

data

A data.frame object with metadata and data columns

var

The column name for your variable of interest

myfactor

The column name for your column containing your factor to group by

normality

The Normality test to be applied, "dagostino" or "shapiro". Default NULL

specifiedNormality

Default NULL leading to non-parametric, can switch by specifying "parametric" or "nonparametric".

correction

Multiple comparison correction argument, default is set at "none"

override

Internal, default 0.05. Set to 0.99 to force pairwise comparison in anova/kw.

returnType

Internal, default is "stats". Alternate "mean", "median". "behemoth" for internal usage.

Value

A data.frame of test results

Examples

File_Location <- system.file("extdata", package = "Coereba")
panelPath <- file.path(File_Location, "ILTPanelTetramer.csv")
binaryPath <- file.path(File_Location, "HeatmapExample.csv")
dataPath <- file.path(File_Location, "ReadyFileExample.csv")
panelData <- read.csv(panelPath, check.names=FALSE)
binaryData <- read.csv(binaryPath, check.names=FALSE)
dataData <- read.csv(dataPath, check.names=FALSE)

All <- Coereba_MarkerExpressions(data=dataData, binary=binaryData,
 panel=panelData, starter="SparkBlue550")
#> Error in Coereba_MarkerExpressions(data = dataData, binary = binaryData,     panel = panelData, starter = "SparkBlue550"): unused arguments (data = dataData, binary = binaryData, panel = panelData, starter = "SparkBlue550")

TheStats <- Utility_Stats(data=All, var="CD62L",
 myfactor="ptype", normality="dagostino")
#> Error: object 'All' not found